// src/content-handlers/iiif/modules/iiif-av-component-module/src/interfaces/component-content.ts
// eslint-disable-next-line @typescript-eslint/interface-name-prefix
export interface IAVComponentContent {
  currentTime: string;
  collapse: string;
  duration: string;
  expand: string;
  mute: string;
  next: string;
  pause: string;
  play: string;
  previous: string;
  unmute: string;
  fastForward: string;
  fastRewind: string;
  noMediaMessage?: string;
}